load feed list from database
